fmt and refactor
